move modals to modal directory
