+++ src/components/modal/NewRequestModal.vue
@@ -4,9 +4,9 @@ import { openModal, closeModal } from 'jenesius-vue-modal'
 import { ref } from 'vue'
 
 import NewTagModal from './NewTagModal.vue'
-import Button from './shared/Button.vue'
-import MarkdownIt from './shared/MarkdownIt.vue'
-import VueSelect from './shared/VueSelect.vue'
+import Button from '/@/components/shared/Button.vue'
+import MarkdownIt from '/@/components/shared/MarkdownIt.vue'
+import VueSelect from '/@/components/shared/VueSelect.vue'
 import type { Request } from '/@/lib/apis'
 import apis from '/@/lib/apis'
 import clubBudgetRequestTemplate from '/@/md/clubBudgetRequest.md?raw'
