+++ src/components/modal/StatusChangeModal.vue
@@ -2,12 +2,12 @@
 import { closeModal } from 'jenesius-vue-modal'
 import { ref } from 'vue'
 
-import type { StatusEnum } from '../lib/apis/generated'
-import { useRequestDetailStore } from '../stores/requestDetail'
-import Button from './shared/Button.vue'
-import MarkdownIt from './shared/MarkdownIt.vue'
-import StatusChip from './shared/StatusChip.vue'
+import Button from '/@/components/shared/Button.vue'
+import MarkdownIt from '/@/components/shared/MarkdownIt.vue'
+import StatusChip from '/@/components/shared/StatusChip.vue'
 import apis from '/@/lib/apis'
+import type { StatusEnum } from '/@/lib/apis'
+import { useRequestDetailStore } from '/@/stores/requestDetail'
 
 type Props = { nextStatus: StatusEnum | '' }
 
@@ -29,7 +29,7 @@ async function putStatus(nextStatus: StatusEnum | '', comment: string) {
 </script>
 
 <template>
-  <div class="pt-8 bg-white">
+  <div class="pt-8 bg-white w-2/5">
     <h1 class="text-3xl text-center">申請の状態変更</h1>
     <div class="flex flex-col gap-4 justify-around mx-12 mt-8 h-4/5">
       <div class="flex items-center">
