+++ src/pages/RequestDetails.vue
@@ -4,9 +4,9 @@ import { onMounted } from 'vue'
 import { useRoute } from 'vue-router'
 
 import NewComment from '/@/components/NewComment.vue'
-import NewTransactionModal from '/@/components/NewTransactionModal.vue'
 import RequestDetail from '/@/components/RequestDetail.vue'
 import RequestLogs from '/@/components/RequestLogs.vue'
+import NewTransactionModal from '/@/components/modal/NewTransactionModal.vue'
 import Button from '/@/components/shared/Button.vue'
 import { useFileStore } from '/@/stores/file'
 import { useRequestDetailStore } from '/@/stores/requestDetail'
