+++ src/pages/Requests.vue
@@ -5,9 +5,9 @@ import { onMounted, ref, watch } from 'vue'
 import { useRoute } from 'vue-router'
 import type { LocationQueryValue } from 'vue-router'
 
-import NewRequestModal from '../components/NewRequestModal.vue'
 import RequestFilteringMenu from '/@/components/RequestFilteringMenu.vue'
 import RequestItem from '/@/components/RequestItem.vue'
+import NewRequestModal from '/@/components/modal/NewRequestModal.vue'
 import Button from '/@/components/shared/Button.vue'
 import PaginationBar from '/@/components/shared/PaginationBar.vue'
 import { useGroupStore } from '/@/stores/group'
